This commit implements the following:
URS G2 elements trimming. The set of URS G2 elements needed by the protocol is determined by the set of maximal degrees of the polynomials that are committed. This set of polynomials is determined by the set of circuits supported by the same instance of the URS. This commit removes all other URS G2 elements.