This commit implements the following: (#2)
* This commit implements the following:

URS G2 elements trimming. The set of URS G2 elements needed by the protocol is determined by the set of maximal degrees of the polynomials that are committed. This set of polynomials is determined by the set of circuits supported by the same instance of the URS. This commit removes all other URS G2 elements.

* check the presense of the necessary G2 elements upon the batch verification

* use algebra's multiexp routine

* use algebra's wNAF methods instead of current implementation