separate out verifier index (#10)
This PR makes a separate verifier index which only contains those values that are needed by the verifier. The one thing that still needs to be done is only taking the subset of G2 elements from the index that are needed for the degree checks, but I figured that could be done after the degree-bounded verifier is implemented.