Modified scalar challenge (#32)
This PR defines an injective map f: 2^128 -> F, and replaces a subset of the challenges `c` in marlin with `f(c)`. 

The reason is that for these challenges, the verifier previously computed `c * P` (for some point `P`), but, it is much cheaper to compute `f(c) * P` inside the circuit. Therefore, one thinks of the value `c` squeezed from the sponge as an encoding of the field element `f(c)`.

Or another way of thinking about it is that instead of sampling challenges as a uniform random 128 bit string, we select a uniformly random element from the set { f(c) | c a 128 bit challenge } and since f is injective this is ok.

The technique (which uses an endomorphism on the curve) is described in the halo paper.