[kimchi][witness] refactor kimchi to make use of the witness
this is a light refactor. There's more places that could benefit from
switching to this type. One small problem that remains to be solved:

the prover function expects the witness to have remaining room for the zk rows,
whereas we now create the witness of the full size of the domain.
I think it is better this way, but we need to think of a way to add zk rows,
as part of this abstraction.