Starting Kimchi MSM circuit
This commit sets the basic structure for the Kimchi MSM circuit which aims to be
a circuit allowing to prove a large MSM with group elements formed from a
foreign field.

It defines a generic prover and verifier over a indexed number of columns called
X_i, and using KZG over the curve BN254.
In following PRs, the prover and verifier will use iterators to absorb
commitments and evaluate points.
Also, MVLookup will be integrated.

Run `cargo run --bin main --release` to verify the verifier accepts a random
proof.