Kimchi/verifier: stop relying on iterator on option
commitments.lookup is an option. By using iter followed by for_each, the reader
might suppose it is a list.
IMO, it seems more readable to check if lookups are used, unwrap the lookup
commitments, without verifying if an error is raised as it
should never happened, and get the commitment to the lookup aggregator.