MSM/MVLookup: add challenges into column environment
The generation of the challenges should be at the same round for both the prover
and the verifier.
We do not have a typed sponge unfortunately to detect that we coin at the right
time. Maybe later.
I am integrating lookups for the serialization subcircuits. We will have tests
when this is done.