Utils: ensuring that the requested number of chunks is returned
As noted in the trait documentation.

When we do Vec::with_capacity(n), it does NOT mean that the vector is going to
be of size n , and not more.
It does only mean that, initially, the vector will point to an allocated
contiguous piece of memory that can handle n elements of the type parameter of
your vector. It might be reallocating if you try to push n + 1 elements.

Source: https://doc.rust-lang.org/std/vec/struct.Vec.html#method.with_capacity.

Therefore, if you want to enforce your vector to be of size n, and not more, you
must verify it yourself.