Folding: remove zkrows and lagrange basis variants from folding expressions
We do not need them at the moment. We might want to add them later, but it is
not used now, and it adds noise to the code.
The folding expression framework should only handle multivariate polynomials
describing relation constraints for now. We can add lookup and others later,
when needed.