o1VM: rename `Trace(r)` into `DecomposableTrace(r)` + add docs
This commit groups multiple changes.
First, it renames Trace and Tracer into `DecomposableTrace` and
`DecomposableTracer`.
These names represent more the type of "objects"/traces the structure and the
trait describe.

On top of that, additional documentation has been added at the top of the file
trace.rs. Links to the structures using square brackets are added to detect
old/inconsistent documentations mentioning code in the CI.

To finish, lines are broken into 80 characters long-lines to respect (old)
coding conventions.