Folding: reorganize testing structure
It was a bit disorganized, with feature enabled in the main crate.
The dependency bn254 is now only activated when we build the tests, which is the
expected behavior.
We also moved the tests into the subdirectory tests, which is the expected
structure in a Cargo project.
Moving into a subdirectory requires to make some changes, like nested the
evaluations into a structure with unamed fields because it is forbidden in Rust
to implement a trait for a list outside of the crate where the trait is defined.

It also shows to an external user how to use the folding library externall.
Until now, as the tests were considered inside the library, it wasn't fully
representing a real external usecase.

To finish, it also simplifies some dependencies flags. The dependency bn254 is
only required in folding when running tests, confirming that the folding library
is generic for every fields/curves.

The reviewer can verify that the tests are run correctly by using:
```
cargo nextest run --release --all-features
```

from the folding subdirectory.