Folding: make the variable name more explicits
ins_i and wit_i variable names are reused. However, their semantics changed when
called compute_extension on the environment. The values contain the
commitments/evaluations added by quadriticization.

In a future PR, the values will be returned as part of the folded output to pass
them to the IVC circuit.