Folding: output the relaxed extended instances
It is useful information to pass to the IVC circuit.
I would suggest to improve the interface later.

Adding also documentation.

Regarding the usage of `clone`, I am not happy with it. It will imply a possible
non-negligeable memory overhead. We should avoid this when the first version is
released.