Folding: simplify the interface and rename fields
The folding interface works as follow:
An original instance is extended using quadriticization.
When the instance has been extended, it is relaxed.

A relaxed instance will therefore contain the extended instance, and the
extended instance will contain the extra columns added by the
quadriticization process.

Before the changes, the interface was using the concept of "inner", and
was defining methods on structures to get reference to different fields.
It does add unecessary code, and also using shared terms between
(semantically different) structures complexifies the understanding.

This commit aims to simplify the reading of the interface, by using
explicit field name. Also, it makes the fields of the structure
accessible publicly, and allows to remove some methods that adds "noise"
to the codebase. References are automatically infered, and therefore
functions returning references are not required.