Poseidon: expose publicly bn254 parameters for 8-56-5-3-2 128 bits
This will make it reusable outside. Also, it is important that the user knows
that it is only for bn254, and no other curves. The parameters must be
different for other curves.

The dependency on bn254 should be later considered as a feature. The IVC circuit
is generic enough over any elliptic curve without cycle.