IVC: prepare code for folding APP and IVC
Mostly cleaning old code, and prepare the witness and instance to include the
IVC columns

In the next PR, the folding env and config will be changed to use the new
generic folding env created some days ago. It will clean the code.

From there, I will continue to concatenate the witnesses and instances