Poly-commitment/KZG: uniform definition of `SRS::create`
And include additional documentation for the method.
KZG won't be used in production in the near future, and there is no plan for
now.
Therefore, it seems reasonable to have these changes, with the warning.

Note that we should get rid of create_trusted_setup in IPA::SRS, and stop
relying on IPA::SRS to define the KZG SRS. It was initially introduced to avoid
to reimplement all methods and have something quick for testing.