Arrabbiata: introduce two kind of "last digest".
The verifier will need to recompute the challenges, and for this, we need to
keep track of the digest of the program state before we committed at the
previous iteration.

For the current iteration, we keep track also of the digest of the program state
we have to start from.