Arrabbiata: introduce the concept of "program" in the environment
The structure Program would encapsulate all the information reg. an
element of the NP relation.
In other words, it contains all the public and private values, including the
challenges. It will also contain the input and the expected output that will be
passed in public values. In addition to that, the accumulations will also be kept.

The environment used by the interpreter will work with two different instances
of the program structure, one for each curve cycle.