plonk-wasm: Fix the warning until we can update
`ark-ff` contains breaking changes so it might take a while to update.
On the current version of Rust, we get a bunch of annoying scoping
warnings that are easily fixed. I've got an open PR to
`arkworks-rs/algebra` (the workspace repo) to backport there changes
from `0.5.x` onto a 0.4.x branch. Hopefully they accept the PR and do a
new patch release. If not, we can update the dependency.